+++ tests/test_coding/test_resources.py
@@ -1,8 +1,8 @@
+import time
 import unittest
 from concurrent.futures import Future
 
 from satella.coding.concurrent import call_in_separate_thread
-
 from satella.coding.resources import CPManager
 
 
@@ -17,14 +17,20 @@ class TestResources(unittest.TestCase):
                 Connection.total_connections += 1
                 self.id = Connection.total_connections
                 self.value_error_emitted = False
+                self.closed = False
 
             def do(self):
+                if self.closed:
+                    raise RuntimeError('Connection closed')
                 if self.value_error_emitted:
                     raise RuntimeError('do called despite raising ValueError earlier')
                 self.i += 1
                 if self.i == 3:
                     self.value_error_emitted = True
                     raise ValueError()
+
+            def close(self):
+                self.closed = True
 
         class InheritCPManager(CPManager):
             def __init__(self, *args):
@@ -36,6 +42,7 @@ class TestResources(unittest.TestCase):
 
             def acquire_connection(self):
                 v = super().acquire_connection()
+                print(f'Acquiring connection {v.id}')
                 if v.id in self.already_acquired:
                     raise RuntimeError('Reacquiring an acquired connection')
                 self.already_acquired.add(v.id)
@@ -45,11 +52,12 @@ class TestResources(unittest.TestCase):
                 super().fail_connection(connection)
 
             def release_connection(self, connection) -> None:
+                print(f'Releasing connection {connection.id}')
                 self.already_acquired.remove(connection.id)
                 super().release_connection(connection)
 
             def teardown_connection(self, connection) -> None:
-                ...
+                connection.close()
 
         cp = InheritCPManager(5, 3)
 
@@ -59,6 +67,7 @@ class TestResources(unittest.TestCase):
         def do_call():
             for _ in range(10):
                 conn = cp.acquire_connection()
+                time.sleep(1)
                 try:
                     conn.do()
                 except ValueError:
@@ -66,9 +75,12 @@ class TestResources(unittest.TestCase):
 
                 cp.release_connection(conn)
 
-        ret = do_call()     # type: Future
+        ret = do_call()  # type: Future
+        ret2 = do_call()
         cp.release_connection(conns.pop())
-        ret.result(timeout=15)
+        cp.release_connection(conns.pop())
+        ret.result(timeout=20)
+        ret2.result(timeout=20)
 
         while conns:
             cp.release_connection(conns.pop())