+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.21.3
 
 * fixed invalidate for CPManager
+* exported max_number for CPManager